+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.4.2 ()
 * WIP
+* Added auto table creation to reduce installation issues.
 
 ## 1.4.1 (12/06/2017)
 * Add CopAddAdmin, CopAdd, CopRem and CopRank rcon command